refactor: 인수테스트 픽스처 수정
