+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/CommentAcceptanceTest.java
@@ -1,7 +1,7 @@
 package com.woowacourse.gongseek.acceptance;
 
 import static com.woowacourse.gongseek.acceptance.support.ArticleFixtures.게시물을_등록한다;
-import static com.woowacourse.gongseek.acceptance.support.AuthFixtures.로그인한다;
+import static com.woowacourse.gongseek.acceptance.support.AuthFixtures.로그인을_한다;
 import static com.woowacourse.gongseek.acceptance.support.CommentFixtures.댓글을_등록한다;
 import static com.woowacourse.gongseek.acceptance.support.CommentFixtures.댓글을_삭제한다;
 import static com.woowacourse.gongseek.acceptance.support.CommentFixtures.댓글을_수정한다;
@@ -18,11 +18,12 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
+@SuppressWarnings("NonAsciiCharacters")
 public class CommentAcceptanceTest extends AcceptanceTest {
 
     @Test
     void 유저가_깃허브로_로그인을_하고_댓글을_등록할_수_있다() {
-        TokenResponse 엑세스토큰 = 로그인한다(주디);
+        TokenResponse 엑세스토큰 = 로그인을_한다(주디);
         ArticleIdResponse 게시글번호 = 게시물을_등록한다(엑세스토큰);
 
         ExtractableResponse<Response> 댓글 = 댓글을_등록한다(엑세스토큰, 게시글번호);
@@ -40,8 +41,9 @@ public class CommentAcceptanceTest extends AcceptanceTest {
 
     @Test
     void 댓글을_조회할_수_있다() {
-        TokenResponse 엑세스토큰 = 로그인한다(주디);
+        TokenResponse 엑세스토큰 = 로그인을_한다(주디);
         ArticleIdResponse 게시글번호 = 게시물을_등록한다(엑세스토큰);
+
         댓글을_등록한다(엑세스토큰, 게시글번호);
 
         List<CommentResponse> 댓글리스트 = 댓글을_조회한다(엑세스토큰, 게시글번호);
@@ -51,7 +53,7 @@ public class CommentAcceptanceTest extends AcceptanceTest {
 
     @Test
     void 댓글을_작성한_유저일_경우_댓글을_수정할_수_있다() {
-        TokenResponse 엑세스토큰 = 로그인한다(주디);
+        TokenResponse 엑세스토큰 = 로그인을_한다(주디);
         ArticleIdResponse 게시글번호 = 게시물을_등록한다(엑세스토큰);
         댓글을_등록한다(엑세스토큰, 게시글번호);
         List<CommentResponse> 댓글리스트 = 댓글을_조회한다(엑세스토큰, 게시글번호);
@@ -63,7 +65,7 @@ public class CommentAcceptanceTest extends AcceptanceTest {
 
     @Test
     void 댓글을_작성한_유저일_경우_댓글을_삭제할_수_있다() {
-        TokenResponse 엑세스토큰 = 로그인한다(주디);
+        TokenResponse 엑세스토큰 = 로그인을_한다(주디);
         ArticleIdResponse 게시글번호 = 게시물을_등록한다(엑세스토큰);
         댓글을_등록한다(엑세스토큰, 게시글번호);
         List<CommentResponse> 댓글리스트 = 댓글을_조회한다(엑세스토큰, 게시글번호);
