+++ backend/src/test/java/com/woowacourse/gongseek/acceptance/support/AuthFixtures.java
@@ -1,15 +1,17 @@
 package com.woowacourse.gongseek.acceptance.support;
 
 import com.woowacourse.gongseek.auth.presentation.dto.OAuthCodeRequest;
+import com.woowacourse.gongseek.auth.presentation.dto.OAuthLoginUrlResponse;
 import com.woowacourse.gongseek.auth.presentation.dto.TokenResponse;
 import com.woowacourse.gongseek.auth.support.GithubClientFixtures;
 import io.restassured.RestAssured;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 
+@SuppressWarnings("NonAsciiCharacters")
 public class AuthFixtures {
 
-    public static TokenResponse 로그인한다(GithubClientFixtures client) {
+    public static TokenResponse 로그인을_한다(GithubClientFixtures client) {
         return RestAssured
                 .given().log().all()
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -20,5 +22,16 @@ public class AuthFixtures {
                 .statusCode(HttpStatus.OK.value())
                 .extract()
                 .as(TokenResponse.class);
+    }
+
+    OAuthLoginUrlResponse 로그인_URL을_얻는다() {
+        return RestAssured
+                .given().log().all()
+                .when()
+                .get("/api/auth/github")
+                .then().log().all()
+                .statusCode(HttpStatus.OK.value())
+                .extract()
+                .as(OAuthLoginUrlResponse.class);
     }
 }
